--- index.html
@@ -1,32 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Team Dream</title>
-    <link rel="stylesheet" href="style.css">
-</head>
-<body>
-<div class="content">
-    <!-- Tab -->
-    <ul class="nav" id="navigation">
-        <li class="nav-item">
-            <a class="tab-link active" href="#" id="defaultOpen" onclick="openTab(event, 'home')">Home</a>
-        </li>
-        <li class="nav-item">
-            <a class="tab-link" href="#" onclick="openTab(event, 'question')">Question</a>
-        </li>
-        <li class="nav-item">
-            <a class="tab-link" href="#" onclick="openTab(event, 'about')">About</a>
-        </li>
-    </ul>
-    <!-- Tab content -->
-    <iframe id="home" class="tab-content" src="home/home.html">
-    </iframe>
-    <iframe id="question" class="tab-content" style="display:none" src="question/question.html">
-    </iframe>
-    <iframe id="about" class="tab-content" style="display:none" src="about/about.html">
-    </iframe>
-</div>
-<script src="index.js"></script>
-</body>
-</html>
