+++ src/about/about.html
@@ -3,13 +3,9 @@
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
-    <link rel="stylesheet" href="about.css">
+    <link rel="stylesheet" href="about.scss">
 </head>
 <body>
-<!--TODO ABOUT UI -->
-<h1>
-    ABOUT
-</h1>
 <main>
     <h1 class="central">СЕРВИС ТЕСТИРОВАНИЯ ЗНАНИЙ</h1>
