+++ src/home/home.html
@@ -3,13 +3,13 @@
 <head>
     <meta charset="UTF-8">
     <title>Title</title>
-    <link rel="stylesheet" href="home_style.css">
+    <link rel="stylesheet" href="home_style.scss">
 </head>
 <body>
 <div class="cards-wrapper">
     <!--USER 1-->
     <div class="card">
-        <img src="img_1.jpg" alt="img1" class="profile-image"/>
+        <img src="../../home/img_1.jpg" alt="img1" class="profile-image"/>
         <span>Full Name</span>
         <h1>USER Name</h1>
         <span>Main Info:</span>
@@ -29,7 +29,7 @@
 
     <!--USER 2-->
     <div class="card">
-        <img src="img_1.jpg" alt="img1" class="profile-image"/>
+        <img src="../../home/img_1.jpg" alt="img1" class="profile-image"/>
         <span>Full Name</span>
         <h1>USER Name</h1>
         <span>Main Info:</span>
@@ -49,7 +49,7 @@
 
     <!--USER 3-->
     <div class="card">
-        <img src="img_1.jpg" alt="img1" class="profile-image"/>
+        <img src="../../home/img_1.jpg" alt="img1" class="profile-image"/>
         <span>Full Name</span>
         <h1>USER Name</h1>
         <span>Main Info:</span>
@@ -69,7 +69,7 @@
 
     <!--USER 4-->
     <div class="card">
-        <img src="img_1.jpg" alt="img1" class="profile-image"/>
+        <img src="../../home/img_1.jpg" alt="img1" class="profile-image"/>
         <span>Full Name</span>
         <h1>USER Name</h1>
         <span>Main Info:</span>
@@ -89,7 +89,7 @@
 
     <!--USER 5-->
     <div class="card">
-        <img src="img_1.jpg" alt="img1" class="profile-image"/>
+        <img src="../../home/img_1.jpg" alt="img1" class="profile-image"/>
         <span>Full Name</span>
         <h1>USER Name</h1>
         <span>Main Info:</span>
