+++ src/home/home_style.scss
@@ -17,10 +17,11 @@
     flex-direction: column;
     align-items: center;
     justify-content: center;
-    margin: 10px;
+    margin: 12px;
     border-radius: 8%;
-    padding: 8px;
+    padding: 10px;
     height: auto;
+    box-shadow: 0px 10px 10px rgba(0, 0, 0, 0.4);
 }
 
 .profile-image{
@@ -34,8 +35,8 @@
 
 .main-info{
     display: flex;
-    flex-flow: column wrap;
     justify-content: flex-start;
+    flex-wrap: wrap;
 }
 
 .card button{
