+++ src/index.html
@@ -1,0 +1,35 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <title>Team Dream</title>
+    <link rel="stylesheet" href="style.scss">
+</head>
+<body>
+<section class="content">
+    <!-- Tab -->
+    <ul class="content__nav" id="navigation">
+        <li class="content__item">
+            <a class="item__link active" href="#" id="defaultOpen" onclick="openTab(event, 'home')">Home</a>
+        </li>
+        <li class="content__item">
+            <a class="item__link" href="#" onclick="openTab(event, 'question')">Question</a>
+        </li>
+        <li class="content__item">
+            <a class="item__link" href="#" onclick="openTab(event, 'about')">About</a>
+        </li>
+    </ul>
+    <!-- Tab content -->
+    <section>
+    <iframe id="home" class="tab-content" src="home/home.html">
+    </iframe>
+    <iframe id="question" class="tab-content" style="display:none" src="question/question.html">
+    </iframe>
+    <iframe id="about" class="tab-content" style="display:none" src="about/about.html">
+    </iframe>
+    </section>
+</section>
+<script type="module" src="index.js"></script>
+</body>
+</html>
