+++ src/style.scss
@@ -3,7 +3,7 @@ body{
 }
 
 .content {
-    height: 500px;
+    height: auto;
     margin: auto;
     max-width: 1170px;
 }
@@ -50,7 +50,7 @@ ul{
     z-index: 2;
     width: 1120px;
     margin-top: -82px;
-    height: 1000px;
+    height: 950px;
     padding: 25px;
     box-shadow: 0px 10px 10px rgba(0, 0, 0, 0.4);
 }
